Fixed linker errors, set the project to only run on intel iGPU for now, Optimus is not playing nice and need to figure out how that works.

+++ SpaceSim/Math/matrix33.cpp
@@ -143,7 +143,7 @@ void Matrix33::minor(int row, int colom, float theminor[])
 	}
 }
 
-float* Matrix33::operator [](unsigned int i)
+float* Matrix33::operator [](size_t i)
 {
 	if (i > 2)
 		return 0;
